requirements.txt
awsebcli==3.19.2
boto3==1.9.191
botocore==1.19.47
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.3
Django==2.1.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
docutils==0.15
future==0.16.0
idna==2.7
jmespath==0.9.4
pathspec==0.5.9
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.3.1
requests==2.20.1
s3transfer==0.2.1
semantic-version==2.5.0
six==1.15.0
stripe==2.32.1
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.7
